[cli] allow to run 'factoryreset' under the diag mode
Before starting a test case, our test script calls the 'factoryreset'
to make the device enters a known state . But the previous test case
may fail to call the 'diag stop' before exiting. This case the next
test case failed to call the 'factoryreset '.

This commit allows the users to call 'factoryreset' under the diag mode.