[spinel] Mapping kMaxSpinelFrame to SPINEL_FRAME_MAX_SIZE in radio_spinel.hpp
The current implementation does not differentiate the max spinel frame size that is used to pack a spinel command and the radio spinel RX frame buffer size which could contain multiple frames.
In radio_spinel_impl.hpp, only one spinel frame can be written when calling function “SendCommand” or “SendReset”.
Therefore the aim of this commit is to allocate, in such a context, only a buffer of one frame instead of a buffer that could contain multiple frames.

Signed-off-by: Gatien Chapon <gatien.chapon@nxp.com>