[spinel] update the updating/adding RCP spinel properity guideline
When adding a new spinel properity or updating the existing spinel
properity, it may cause the host can't work with the old RCP. The
host can be updated independently on some of our products. If the
compatibility issue happens, it causes the host dies after the
host is updated.

To reduce the compatibility issues between the host and RCP, this
commit updates the guideline of updating/adding RCP spinel properities.