[api] allow CSL receiver to gather parent CSL capabilities
This commit extends the Thread API to allow a CSL Receiver application
to adjust its CSL parameters depending on the parent capabilities.

Specifically, it might decide to switch to polling operation instead of
CSL syncrhonization when the attached parent does not support CSL Transmitter
role (Thread Version 2) or it advertises poor CSL accuracy or uncertainty.