[message] extend the message buffer size and its checks
With changes in #7554 the `Metadata` stuct size in `Message` increased,
leading to unsufficient message buffer size when the maximum number
of children is configured, due to the variable size of `ChildMask`.

Furthermore, enabling CoAP Blockwise option also extends the message
buffer size requirements.

Since building with the maximum number of children configured is not
covered for 32-bit architectures, this commit also extends the `check-size`
test using nRF52840 platform to do so.